`extend` scope object instead of messing with Haml::Helpers directly
In order to add translation methods to Haml templates, we were previously using
`include` on the Haml::Helpers module. This has the downside of being an
undoable operation and thus makes testing harder.

Instead, we mangle the code generated by the Haml compiler to include a call to
`extend` the scope object with our module with translation methods, depending
on the selected backend.