Port to Haml 4.0 API
Haml internals has been refactored with version 4.0. The engine has been split
and uses dedicated Parser and Compiler classes. The good news is that they
can be specified in the engine options.

So instead of hacking methods and classes, we now can extend Haml Parser and
Compiler classes to implement translation and catalog facilities.

Support for maruku filters has been split to a new haml-contrib gem. So we add
that to development dependencies in order to run our tests.